+++ lib/src/xml_events/converters/subtree_selector.dart
@@ -10,9 +10,9 @@ import '../events/start_element_event.dart';
 import 'list_converter.dart';
 
 extension XmlSubtreeSelectorExtension on Stream<List<XmlEvent>> {
-  /// From a sequence of [XmlEvent] objects filter out only the event sequences
-  /// of sub-trees for which [predicate] returns `true`.
-  Stream<List<XmlEvent>> selectSubtree(
+  /// From a sequence of [XmlEvent] objects filter the event sequences that
+  /// form sub-trees for which [predicate] returns `true`.
+  Stream<List<XmlEvent>> selectSubtreeEvents(
           Predicate<XmlStartElementEvent> predicate) =>
       transform(XmlSubtreeSelector(predicate));
 }